Update speed parameter

+++ src/Aligner.h
@@ -57,7 +57,7 @@ struct AlignerParams
 	long long colinearGap;
 	long long colinearSplitLen;
 	long long colinearSplitGap;
-	long long speed;
+	double speed;
 	bool fastMode;
 	
 };
